Move duckdb index and dedupe into new integration package
