+++ zavod/zavod/cli.py
@@ -15,9 +15,9 @@ from zavod.crawl import crawl_dataset
 from zavod.store import get_store
 from zavod.archive import clear_data_path, dataset_state_path
 from zavod.exporters import export_dataset
-from zavod.dedupe import get_resolver, get_dataset_linker
-from zavod.dedupe import blocking_xref, merge_entities
-from zavod.dedupe import explode_cluster
+from zavod.integration import get_resolver, get_dataset_linker
+from zavod.integration.dedupe import blocking_xref, merge_entities
+from zavod.integration.dedupe import explode_cluster
 from zavod.runtime.versions import make_version
 from zavod.publish import publish_dataset, publish_failure
 from zavod.tools.load_db import load_dataset_to_db
