+++ zavod/zavod/integration/__init__.py
@@ -1,0 +1,3 @@
+from zavod.integration.dedupe import get_resolver, get_dataset_linker
+
+__all__ = ["get_resolver", "get_dataset_linker"]
