+++ zavod/zavod/integration/duckdb_index.py
