+++ zavod/zavod/tests/exporters/test_metadata.py
@@ -2,7 +2,7 @@ import json
 from zavod import settings
 from zavod.meta import Dataset
 from zavod.store import get_store
-from zavod.dedupe import get_resolver
+from zavod.integration import get_resolver
 from zavod.crawl import crawl_dataset
 from zavod.exporters import export_dataset
