+++ zavod/zavod/tests/fixtures/testdataset_dedupe/crawler.py
@@ -1,4 +1,3 @@
-
 import csv
 from pathlib import Path
 from zavod.context import Context
@@ -6,13 +5,13 @@ from zavod.context import Context
 
 LOCAL_PATH = Path(__file__).parent / "dataset.csv"
 
+
 def crawl(context: Context):
 
     data_path = context.get_resource_path("source.csv")
     with open(LOCAL_PATH, "r") as fh:
         with open(data_path, "w") as out:
             out.write(fh.read())
-    
 
     with open(data_path, "r") as fh:
         for row in csv.DictReader(fh):
