+++ zavod/zavod/tests/tools/test_dump_file.py
@@ -2,7 +2,7 @@ from nomenklatura.judgement import Judgement
 from nomenklatura.statement import Statement, CSV, read_path_statements
 
 from zavod.meta import Dataset
-from zavod.dedupe import get_resolver
+from zavod.integration import get_resolver
 from zavod.crawl import crawl_dataset
 from zavod.tools.dump_file import dump_dataset_to_file
 from zavod.archive import iter_dataset_statements, dataset_state_path
