+++ zavod/zavod/tests/tools/test_summarize.py
@@ -1,6 +1,6 @@
 from zavod.tools.summarize import summarize
 from zavod.store import get_store
-from zavod.dedupe import get_resolver
+from zavod.integration import get_resolver
 from zavod.meta import Dataset
 from zavod.crawl import crawl_dataset
